# `FastGuid.NewGuid()` helper