--- a/src/Autofac/Core/UniqueService.cs
+++ b/src/Autofac/Core/UniqueService.cs
@@ -1,6 +1,8 @@
 ﻿// Copyright (c) Autofac Project. All rights reserved.
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
+using Autofac.Util;
+
 namespace Autofac.Core;
 
 /// <summary>
@@ -14,7 +16,7 @@ public sealed class UniqueService : Service
     /// Initializes a new instance of the <see cref="UniqueService"/> class.
     /// </summary>
     public UniqueService()
-        : this(Guid.NewGuid())
+        : this(FastGuid.NewGuid())
     {
     }