--- a/src/Autofac/Features/Collections/CollectionRegistrationSource.cs
+++ b/src/Autofac/Features/Collections/CollectionRegistrationSource.cs
@@ -143,7 +143,7 @@ internal class CollectionRegistrationSource : IRegistrationSource, IPerScopeRegi
             });
 
         var registration = new ComponentRegistration(
-            Guid.NewGuid(),
+            FastGuid.NewGuid(),
             activator,
             CurrentScopeLifetime.Instance,
             InstanceSharing.None,