It is well-known that standard `Guid.NewGuid()` is slow.
And even slower on Linux https://github.com/dotnet/runtime/issues/13628

If GUIDs are not published outside of process it is enough to use global 64-bit counter as part of GUID generator.