@alistairjevans 2024-03-15
Thanks, interesting suggestion; can you please run our benchmarks and post before and after results here?