@botinko 2024-03-16
Benchmark: 

```
| Method  | Mean      | Error     | StdDev    |
|-------- |----------:|----------:|----------:|
| Default | 34.354 ns | 0.2249 ns | 0.2104 ns |
| Fast    |  8.337 ns | 0.0624 ns | 0.0553 ns |

```

```
    [Benchmark]
    public Guid Default() => Guid.NewGuid();

    [Benchmark]
    public Guid Fast() => FastGuid.NewGuid();
```

For different OS it could be different. In mentioned dotnet runtime issue they say ~900ns on Linux. On Linux `Guid.NewGuid()` uses `/dev/urandom` as a source for randomness. `/dev/urandom`  serve as cryptographically secure pseudorandom number generators. They allow access to a CSPRNG that is seeded with entropy from environmental noise, collected from device drivers and other sources. It's known fact, that it have limited trougput, especially considering that in Kubernetes many containers share one physical machine and possible same `/dev/urandom` stream.