@alistairjevans 2024-03-16
Ok, can you compare the existing Autofac benchmarks as well, the ones in our "bench" folder (https://github.com/autofac/Autofac/tree/develop/bench/Autofac.Benchmarks)?

Want to understand the overall impact of the change on Autofac, based on how many guids we generate.