@SergeiPavlov 2024-03-18
What are specific Benchmarks which depends on `.NewGuid()`

I ran the entire set Before & After this change and don't see performance improvement
But benchmarks mostly test `Resolve*()` methods, not registrations where GUID may be generated