@SergeiPavlov 2024-03-18
I have tried to mix bits of `Guid`:
```
    public static Guid NewGuid()
    {
        var v = Interlocked.Increment(ref variablePart);
        var h = (int)v * 521;               // one-to-one function to improve Guid's hash distribution
        return new(h, (short)(v >> 32), (short)(v >> 48), BasePart);
    }
```

But it does not show visible improvement.
`Guid.GetHashCode()` XORs all 32-bit parts of Guid, so the distribution should be uniform