@SergeiPavlov 2024-03-18
Thanks to @snaumenko-st for further improvement of this PR: https://github.com/autofac/Autofac/pull/1414/commits/88272de910e152b983e5f684c5e16b1beb39e0d8

Benchmark results look not worse than before.