@tillig 2024-04-05
Trying to figure out the state of this so I can get a 0.1.0 release out for #1415. I may just release that one without this, regardless.

However, looking at the chain here, it sounds like:

- The first run of benchmarks showed that things actually got _worse_.
- Subsequent runs (not shown) indicate things aren't worse, but also aren't any better.
- The primary use case here is to improve _registration_ time, which happens effectively once per application lifecycle.

It's not interesting to add an optimization that doesn't do anything, so I'm not sure where that leaves us.

If it's something where we can show some _really marked improvement_ like "it used to take 10 seconds to start this app and now it takes half a second" or whatever, I'm all for it. But I'm not super interested in optimizing the "we told Autofac to scan and register every type in the entire app and all assemblies" use case - folks shouldn't be doing that, and the pain may be a forcing factor to tell them they're doing something wrong, especially if we're still supporting `AnyConcreteTypeNotAlreadyRegisteredSource`.