@SergeiPavlov 2024-04-05
> * The primary use case here is to improve _registration_ time, which happens effectively once per application lifecycle.

In our case most of restrations happen at App start, but a few additional happen at every Web Request.
That is why we are interested in this improvement.