@SergeiPavlov 2024-04-05
Added a `RegistrationBenchmark`:

Before:
| Method   | Mean     | Error    | StdDev   | Gen0   | Gen1   | Allocated |
|--------- |---------:|---------:|---------:|-------:|-------:|----------:|
| Register | 53.76 us | 0.656 us | 0.548 us | 0.3052 | 0.2441 |  26.77 KB |

After:
| Method   | Mean     | Error    | StdDev   | Gen0   | Gen1   | Allocated |
|--------- |---------:|---------:|---------:|-------:|-------:|----------:|
| Register | 49.55 us | 0.648 us | 0.575 us | 0.3052 | 0.2441 |  27.82 KB |