@tillig 2024-04-05
Let me verify I'm reading this right:

After adding the fast GUID optimization, across six registrations you'll save about four microseconds (0.7us per registration) but allocate 1KB more memory to get that.

Does that sound right?