@SergeiPavlov 2024-04-05
Yes.

Second run of this benchmark confirms.
Probably new Guids have not so random distribution and Dictionary allocates more buckets for them.