@SergeiPavlov 2024-04-06
We see following hotspot on `NewGuid()` during profiling
![image](https://github.com/autofac/Autofac/assets/1176609/c4f1fd71-c5ab-476b-82dd-3b10a730258e)

So it looks not-insignificant