@alistairjevans 2024-04-06
Was that profiling session captured on windows or Linux?