@SergeiPavlov 2024-04-06
Linux Debian x64 .NET8