@SergeiPavlov 2024-04-07
> Ah; is that also where you've been running benchmarks?

Not exactly
I was running benchmerks on Ubuntu x64 on real PC
but the App profiling was under Docker container in the Azure cloud

May be `.NewGuid()` implementation differs in that environment