@SergeiPavlov 2024-04-09
> Is it possible to run the benchmarks in that environment?

I did. The results on Azure are the same as on bare Linux.
`FastGuid` ~10x times faster than `Guid.NewGuid()`

A agree, it is not enough to be merge into Autofac